# code-template-master/src/main/java/com/changgou/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "微服务工程代码生成器,生成Dao、Service、Controller、Pojo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "深圳黑马训练营代码生成器",
    "termsOfService": "https://github.com/shenkunlin/code-template.git",
    "contact": {
      "email": "shenkunlin@itcast.cn"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/album/{id}": {
      "delete": {
        "tags": [
          "AlbumController"
        ],
        "summary": "鏍规嵁ID鍒犻櫎Album",
        "description": "鏍规嵁ID鍒犻櫎Album鏂规硶璇︽儏",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "鏍规嵁ID鍒犻櫎Album鏂规硶璇︽儏",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鏍规嵁ID鍒犻櫎Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "AlbumController"
        ],
        "summary": "鏍规嵁ID淇敼Album",
        "description": "鏍规嵁ID淇敼Album鏂规硶璇︽儏",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "鏍规嵁ID淇敼Album鏂规硶璇︽儏",
            "required": true
          },
          {
            "in": "body",
            "name": "Album",
            "description": "浼犲叆Album鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Album"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鏍规嵁ID淇敼Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "AlbumController"
        ],
        "summary": "鏍规嵁ID鏌ヨAlbum",
        "description": "鏍规嵁ID鏌ヨAlbum鏂规硶璇︽儏",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "鏍规嵁ID淇敼Album鏂规硶璇︽儏",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鏍规嵁ID淇敼Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦Album禄"
            }
          }
        }
      }
    },
    "/album/search": {
      "post": {
        "tags": [
          "AlbumController"
        ],
        "summary": "涓嶅甫鍒嗛〉鏉′欢鎼滅储Album",
        "description": "涓嶅甫鍒嗛〉鏉′欢鎼滅储Album鏂规硶璇︽儏",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Album",
            "description": "浼犲叆Album鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Album"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "涓嶅甫鍒嗛〉鎼滅储Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦Album禄禄"
            }
          }
        }
      }
    },
    "/album/search/{page}/{size}": {
      "get": {
        "tags": [
          "AlbumController"
        ],
        "summary": "鍒嗛〉鍒楄〃鏌ヨAlbum",
        "description": "鍒嗛〉鍒楄〃鏌ヨAlbum鏂规硶璇︽儏",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "褰撳墠椤�",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "姣忛〉鏄剧ず鏉℃暟",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "涓嶅甫鍒嗛〉鎼滅储Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦Album禄禄"
            }
          }
        }
      },
      "post": {
        "tags": [
          "AlbumController"
        ],
        "summary": "鍒嗛〉鏉′欢鎼滅储Album",
        "description": "鍒嗛〉鏉′欢鎼滅储Album鏂规硶璇︽儏",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "褰撳墠椤�",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "姣忛〉鏄剧ず鏉℃暟",
            "required": true
          },
          {
            "in": "body",
            "name": "Album",
            "description": "浼犲叆Album鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Album"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鍒嗛〉鏉′欢鎼滅储Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦Album禄禄"
            }
          }
        }
      }
    },
...
    "/brand/{id}": {
      "delete": {
        "tags": [
          "BrandController"
        ],
        "summary": "鏍规嵁ID鍒犻櫎Brand",
        "description": "鏍规嵁ID鍒犻櫎Brand鏂规硶璇︽儏",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "鏍规嵁ID鍒犻櫎Brand鏂规硶璇︽儏",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鏍规嵁ID鍒犻櫎Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "BrandController"
        ],
        "summary": "鏍规嵁ID淇敼Brand",
        "description": "鏍规嵁ID淇敼Brand鏂规硶璇︽儏",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "鏍规嵁ID淇敼Brand鏂规硶璇︽儏",
            "required": true
          },
          {
            "in": "body",
            "name": "Brand",
            "description": "浼犲叆Brand鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Brand"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鏍规嵁ID淇敼Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "BrandController"
        ],
        "summary": "鏍规嵁ID鏌ヨBrand",
        "description": "鏍规嵁ID鏌ヨBrand鏂规硶璇︽儏",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "鏍规嵁ID淇敼Brand鏂规硶璇︽儏",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鏍规嵁ID淇敼Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦Brand禄"
            }
          }
        }
      }
    },
    "/brand/search": {
      "post": {
        "tags": [
          "BrandController"
        ],
        "summary": "涓嶅甫鍒嗛〉鏉′欢鎼滅储Brand",
        "description": "涓嶅甫鍒嗛〉鏉′欢鎼滅储Brand鏂规硶璇︽儏",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Brand",
            "description": "浼犲叆Brand鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Brand"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "涓嶅甫鍒嗛〉鎼滅储Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦Brand禄禄"
            }
          }
        }
      }
    },
    "/brand/search/{page}/{size}": {
      "get": {
        "tags": [
          "BrandController"
        ],
        "summary": "鍒嗛〉鍒楄〃鏌ヨBrand",
        "description": "鍒嗛〉鍒楄〃鏌ヨBrand鏂规硶璇︽儏",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "褰撳墠椤�",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "姣忛〉鏄剧ず鏉℃暟",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "涓嶅甫鍒嗛〉鎼滅储Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦Brand禄禄"
            }
          }
        }
      },
      "post": {
        "tags": [
          "BrandController"
        ],
        "summary": "鍒嗛〉鏉′欢鎼滅储Brand",
        "description": "鍒嗛〉鏉′欢鎼滅储Brand鏂规硶璇︽儏",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "褰撳墠椤�",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "姣忛〉鏄剧ず鏉℃暟",
            "required": true
          },
          {
            "in": "body",
            "name": "Brand",
            "description": "浼犲叆Brand鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Brand"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鍒嗛〉鏉′欢鎼滅储Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦Brand禄禄"
            }
          }
        }
      }
    },
...
    "/category/{id}": {
      "delete": {
        "tags": [
          "CategoryController"
        ],
        "summary": "鏍规嵁ID鍒犻櫎Category",
        "description": "鏍规嵁ID鍒犻櫎Category鏂规硶璇︽儏",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "鏍规嵁ID鍒犻櫎Category鏂规硶璇︽儏",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鏍规嵁ID鍒犻櫎Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "CategoryController"
        ],
        "summary": "鏍规嵁ID淇敼Category",
        "description": "鏍规嵁ID淇敼Category鏂规硶璇︽儏",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "鏍规嵁ID淇敼Category鏂规硶璇︽儏",
            "required": true
          },
          {
            "in": "body",
            "name": "Category",
            "description": "浼犲叆Category鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Category"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鏍规嵁ID淇敼Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "CategoryController"
        ],
        "summary": "鏍规嵁ID鏌ヨCategory",
        "description": "鏍规嵁ID鏌ヨCategory鏂规硶璇︽儏",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "鏍规嵁ID淇敼Category鏂规硶璇︽儏",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鏍规嵁ID淇敼Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦Category禄"
            }
          }
        }
      }
    },
    "/category/search": {
      "post": {
        "tags": [
          "CategoryController"
        ],
        "summary": "涓嶅甫鍒嗛〉鏉′欢鎼滅储Category",
        "description": "涓嶅甫鍒嗛〉鏉′欢鎼滅储Category鏂规硶璇︽儏",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Category",
            "description": "浼犲叆Category鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Category"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "涓嶅甫鍒嗛〉鎼滅储Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦Category禄禄"
            }
          }
        }
      }
    },
    "/category/search/{page}/{size}": {
      "get": {
        "tags": [
          "CategoryController"
        ],
        "summary": "鍒嗛〉鍒楄〃鏌ヨCategory",
        "description": "鍒嗛〉鍒楄〃鏌ヨCategory鏂规硶璇︽儏",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "褰撳墠椤�",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "姣忛〉鏄剧ず鏉℃暟",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "涓嶅甫鍒嗛〉鎼滅储Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦Category禄禄"
            }
          }
        }
      },
      "post": {
        "tags": [
          "CategoryController"
        ],
        "summary": "鍒嗛〉鏉′欢鎼滅储Category",
        "description": "鍒嗛〉鏉′欢鎼滅储Category鏂规硶璇︽儏",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "褰撳墠椤�",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "姣忛〉鏄剧ず鏉℃暟",
            "required": true
          },
          {
            "in": "body",
            "name": "Category",
            "description": "浼犲叆Category鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Category"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "鍒嗛〉鏉′欢鎼滅储Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦Category禄禄"
            }
          }
        }
      }
    },
...
    "/categoryBrand/search": {
      "post": {
        "tags": [
          "CategoryBrandController"
        ],
        "summary": "涓嶅甫鍒嗛〉鏉′欢鎼滅储CategoryBrand",
        "description": "涓嶅甫鍒嗛〉鏉′欢鎼滅储CategoryBrand鏂规硶璇︽儏",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "CategoryBrand",
            "description": "浼犲叆CategoryBrand鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/CategoryBrand"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "涓嶅甫鍒嗛〉鎼滅储CategoryBrand"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦CategoryBrand禄禄"
            }
          }
        }
      }
    },
...
    "/para/search": {
      "post": {
        "tags": [
          "ParaController"
        ],
        "summary": "涓嶅甫鍒嗛〉鏉′欢鎼滅储Para",
        "description": "涓嶅甫鍒嗛〉鏉′欢鎼滅储Para鏂规硶璇︽儏",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Para",
            "description": "浼犲叆Para鐨凧SON瀵硅薄",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Para"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(鏃犳晥鐨勭姸鎬佸��)"
                      },
          "403": {
            "description": "403 Forbidden(璇锋眰琚嫆缁�)"
                      },
          "404": {
            "description": "not found(娌℃湁鎵惧埌鐩稿叧璧勬簮)"
                      },
          "405": {
            "description": "Invalid input(鏃犳晥鐨勮緭鍏�)"
                      },
          "500": {
            "description": "鏈嶅姟鍣ㄥ唴閮ㄩ敊璇�"
                      },
          "200": {
            "description": "涓嶅甫鍒嗛〉鎼滅储Para"
            ,
            "schema": {
              "$ref": "#/definitions/Result芦List芦Para禄禄"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Result": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
    "Album": {
      "type": "object",
      "properties": {
        "id": {
          "type": "long",
          "description": "编号"
        },
        "title": {
          "type": "string",
          "description": "相册名称"
        },
        "image": {
          "type": "string",
          "description": "相册封面"
        },
        "imageItems": {
          "type": "string",
          "description": "图片列表"
        }
      },
      "description": "Album"
    },
...
    "Brand": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "品牌id"
        },
        "name": {
          "type": "string",
          "description": "品牌名称"
        },
        "image": {
          "type": "string",
          "description": "品牌图片地址"
        },
        "letter": {
          "type": "string",
          "description": "品牌的首字母"
        },
        "seq": {
          "type": "integer",
          "format": "int32",
          "description": "排序"
        }
      },
      "description": "Brand"
    },
...
  },
...
}